+++ transforms/simplex/softmax-augmented/dirichlet_symmetric.stan
@@ -1,6 +1,16 @@
+functions{
+    real f_lp(int eval_model, vector x, vector alpha)
+    {
+        if (eval_model==1)
+            return dirichlet_lpdf(x | alpha);
+        else
+            return 0; 
+            }
+}
 data {
  int<lower=0> N;
  vector<lower=0>[N] alpha;
+ int eval_model;
 }
 parameters {
  vector[N] y;
@@ -11,5 +21,5 @@ transformed parameters {
 }
 model {
  target += sum(y) - exp(0.5 * logr) / 0.5;
- target += dirichlet_lupdf(x | alpha);
+ target += f_lp(eval_model, x, alpha);
 }
